+++ example/storybook/src/components/Others/Fab/Fab.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Wrapper from '../../Wrapper';
-import { HamburgerIcon, Fab, Icon, Box } from '../../../ui-components';
+import { AddIcon, Fab, Icon, Box } from '../../../ui-components';
 export const FabStory = ({ placement, showLabel, showIcon, ...props }: any) => {
   return (
     <Wrapper>
@@ -12,7 +12,7 @@ export const FabStory = ({ placement, showLabel, showIcon, ...props }: any) => {
         sx={{ _web: { w: 300, h: 300 } }}
       >
         <Fab placement={placement} {...props}>
-          {showIcon && <Icon as={HamburgerIcon} />}
+          {showIcon && <Icon mr="$1" as={AddIcon} />}
           {showLabel && <Fab.Label>Menu</Fab.Label>}
         </Fab>
       </Box>
@@ -20,4 +20,4 @@ export const FabStory = ({ placement, showLabel, showIcon, ...props }: any) => {
   );
 };
 
-export { Fab, HamburgerIcon };
+export { Fab, AddIcon, Icon, Box };