+++ example/storybook/src/components/Overlay/Menu/Menu.stories.tsx
@@ -1,16 +1,8 @@
 import type { ComponentMeta } from '@storybook/react-native';
 import { MenuStory as Menu } from './Menu';
-import { MenuGroupStory as Group } from './MenuGroup';
-import { ComplexMenuGroup } from './ComplexMenuGroup';
-
-// var st = document.createElement('style');
-// var st2 = document.createElement('style');
-
-// st.innerHTML = `#story--menu--basic{ height: 400px }`;
-// st2.innerHTML = `#story--menu--grouped{ height: 400px }`;
-
-// document.body.append(st);
-// document.body.append(st2);
+// import { MenuStory as Menu } from './MenuDefault';
+// import { MenuStory as Menu } from './MyMenuNative';
+// import { MenuStory as Menu } from './FreshMenu';
 
 const MenuMeta: ComponentMeta<typeof Menu> = {
   title: 'stories/OVERLAY/Menu',
@@ -20,17 +12,17 @@ const MenuMeta: ComponentMeta<typeof Menu> = {
       control: 'select',
       options: [
         'bottom',
-        'bottom-end',
-        'bottom-start',
+        'bottom end',
+        'bottom start',
         'top',
-        'top-end',
-        'top-start',
+        'top end',
+        'top start',
         'left',
-        'left-end',
-        'left-start',
+        'left end',
+        'left start',
         'right',
-        'right-end',
-        'right-start',
+        'right end',
+        'right start',
       ],
     },
   },
@@ -40,7 +32,4 @@ const MenuMeta: ComponentMeta<typeof Menu> = {
 };
 
 export default MenuMeta;
-
 export { Menu };
-export { Group };
-export { ComplexMenuGroup };