+++ example/storybook/src/ui-components/Button/styled-components/Root.tsx
@@ -4,33 +4,35 @@ import { Pressable } from 'react-native';
 export default styled(
   Pressable,
   {
-    'borderRadius': '$lg',
+    'borderRadius': '$sm',
     'backgroundColor': '$primary500',
     'flexDirection': 'row',
     'justifyContent': 'center',
     'alignItems': 'center',
 
     '_text': {
-      color: '$textLight50',
-      fontWeight: '$normal',
+      color: '$textLight0',
+      fontWeight: '$semibold',
       _dark: {
-        color: '$textDark50',
+        color: '$textDark0',
       },
     },
 
     '_spinner': {
       props: {
-        color: '$textLight50',
+        color: '$backgroundLight0',
       },
       _dark: {
-        props: { color: '$textDark50' },
+        props: {
+          color: '$backgroundDark0',
+        },
       },
     },
 
     '_icon': {
-      color: '$backgroundLight800',
+      color: '$backgroundLight0',
       _dark: {
-        color: '$backgroundDark50',
+        color: '$backgroundDark0',
       },
     },
 
@@ -65,10 +67,14 @@ export default styled(
               color: '$primary600',
             },
             ':hover': {
-              props: { color: '$primary600' },
+              props: {
+                color: '$primary600',
+              },
             },
             ':active': {
-              props: { color: '$primary700' },
+              props: {
+                color: '$primary700',
+              },
             },
           },
 
@@ -247,9 +253,6 @@ export default styled(
               _web: {
                 boxShadow: 'offset 0 0 0 2px $info400',
               },
-
-              // borderWidth: '$2',
-              // borderColor: '$red900',
             },
           },
         },
@@ -324,6 +327,25 @@ export default styled(
             },
           },
         },
+
+        default: {
+          'bg': '$transparent',
+          ':hover': {
+            bg: '$backgroundLight50',
+          },
+          ':active': {
+            bg: 'transparent',
+          },
+          '_dark': {
+            'bg': 'transparent',
+            ':hover': {
+              bg: '$backgroundDark900',
+            },
+            ':active': {
+              bg: 'transparent',
+            },
+          },
+        },
       },
 
       variant: {
@@ -347,23 +369,23 @@ export default styled(
 
       size: {
         xs: {
-          px: '$3',
-          py: '$2',
+          px: '$3.5',
+          h: '$8',
           _icon: {
-            h: 12,
-            w: 12,
+            h: '$3',
+            w: '$3',
           },
           _text: {
             fontSize: '$xs',
-            lineHeight: '$xs',
+            lineHeight: '$sm',
           },
         },
         sm: {
-          px: '$3.5',
-          py: '$2',
+          px: '$4',
+          h: '$9',
           _icon: {
-            h: 16,
-            w: 16,
+            h: '$4',
+            w: '$4',
           },
           _text: {
             fontSize: '$sm',
@@ -371,11 +393,11 @@ export default styled(
           },
         },
         md: {
-          px: '$4',
-          py: '$2',
+          px: '$5',
+          h: '$10',
           _icon: {
-            h: 18,
-            w: 18,
+            h: '$4.5',
+            w: '$4.5',
           },
           _text: {
             fontSize: '$md',
@@ -383,15 +405,27 @@ export default styled(
           },
         },
         lg: {
-          px: '$5',
-          py: '$2',
+          px: '$6',
+          h: '$11',
           _icon: {
-            h: 18,
-            w: 18,
+            h: '$4.5',
+            w: '$4.5',
           },
           _text: {
             fontSize: '$lg',
-            lineHeight: '$lg',
+            lineHeight: '$xl',
+          },
+        },
+        xl: {
+          px: '$7',
+          h: '$12',
+          _icon: {
+            h: '$5',
+            w: '$5',
+          },
+          _text: {
+            fontSize: '$xl',
+            lineHeight: '$xl',
           },
         },
       },
@@ -759,7 +793,7 @@ export default styled(
 
     '_web': {
       ':focusVisible': {
-        outlineWidth: 2,
+        outlineWidth: '$0.5',
         outlineColor: '$primary700',
         outlineStyle: 'solid',
         _dark: {
@@ -769,7 +803,7 @@ export default styled(
     },
 
     ':disabled': {
-      opacity: 0.5,
+      opacity: 0.4,
     },
   },
   {