+++ example/storybook/src/ui-components/Divider/index.tsx
@@ -1,2 +1,3 @@
-import { Divider } from './styled-components';
-export { Divider };
+import Root from './styled-components/Root';
+import { createDivider } from '@gluestack-ui/divider';
+export const Divider = createDivider({ Root });