+++ example/storybook/src/ui-components/Input/styled-components/Icon.tsx
@@ -7,7 +7,6 @@ export default styled(
     justifyContent: 'center',
     alignItems: 'center',
     _web: {
-      'cursor': 'pointer',
       ':disabled': {
         cursor: 'not-allowed',
       },