+++ packages/button/src/index.tsx
@@ -3,15 +3,16 @@ import { Button as ButtonMain } from './Button';
 import { ButtonGroup } from './ButtonGroup';
 import { ButtonText } from './ButtonText';
 import { ButtonSpinner } from './ButtonSpinner';
+import { ButtonIcon } from './ButtonIcon';
 import type { IButtonComponentType } from './types';
-
 export function createButton<
   ButtonProps,
   TextProps,
   GroupProps,
   GroupHSpacerProps,
   GroupVSpacerProps,
-  SpinnerProps
+  SpinnerProps,
+  IconProps
 >({
   Root,
   Text,
@@ -19,6 +20,7 @@ export function createButton<
   GroupHSpacer,
   GroupVSpacer,
   Spinner,
+  Icon,
 }: {
   Root: React.ComponentType<ButtonProps>;
   Text: React.ComponentType<TextProps>;
@@ -26,16 +28,19 @@ export function createButton<
   GroupHSpacer: React.ComponentType<GroupHSpacerProps>;
   GroupVSpacer: React.ComponentType<GroupVSpacerProps>;
   Spinner: React.ComponentType<SpinnerProps>;
+  Icon: React.ComponentType<IconProps>;
 }) {
   const Button = ButtonMain(Root) as any;
   Button.Text = ButtonText(Text);
   Button.Group = ButtonGroup(Group, GroupHSpacer, GroupVSpacer);
   Button.Spinner = ButtonSpinner(Spinner);
+  Button.Icon = ButtonIcon(Icon);
 
   Button.displayName = 'Button';
   Button.Text.displayName = 'Button.Text';
   Button.Group.displayName = 'Button.Group';
   Button.Spinner.displayName = 'Button.Spinner';
+  Button.Icon.displayName = 'Button.Icon';
 
   return Button as IButtonComponentType<
     ButtonProps,
@@ -43,6 +48,7 @@ export function createButton<
     GroupHSpacerProps,
     GroupVSpacerProps,
     SpinnerProps,
-    TextProps
+    TextProps,
+    IconProps
   >;
 }