+++ packages/vstack/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gluestack-ui/vstack
 
+## 0.1.3
+
+### Patch Changes
+
+- typing fixes
+
+## 0.1.2
+
+### Patch Changes
+
+- fix: spacer issue
+
 ## 0.1.1
 
 ### Patch Changes